eslint fix and refactoring
